requirements.txt
absl-py==0.8.1
aiohttp==3.6.2
astor==0.8.0
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.3.0
beautifulsoup4==4.8.2
box2d-py==2.3.8
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.0
dm-tree==0.1.5
filelock==3.0.12
funcsigs==1.0.2
future==0.18.1
gast==0.2.2
google==2.0.3
google-pasta==0.1.8
GPUtil==1.4.0
grpcio==1.24.3
gym==0.15.3
h5py==2.10.0
idna==2.8
imageio==2.5.0
importlib-metadata==0.23
jsonschema==3.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lz4==2.2.1
Markdown==3.1.1
matplotlib==3.1.1
more-itertools==7.2.0
multidict==4.7.4
networkx==2.3
numpy==1.16.2
numpy-stl==2.10.1
oauthlib==2.1.0
opencv-contrib-python==4.2.0.32
opencv-python==4.1.1.26
opencv-python-headless==4.1.2.30
packaging==19.2
pandas==0.25.1
Pillow==5.4.1
pluggy==0.13.1
protobuf==3.11.2
psutil==5.6.7
py==1.8.0
py-spy==0.3.2
pycparser==2.19
pygame==1.9.6
pyglet==1.3.2
pymunk==5.5.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.3.1
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2019.3
PyWavelets==1.0.3
PyYAML==3.13
ray==0.9.0.dev0
redis==3.3.11
requests==2.22.0
scikit-image==0.16.2
scipy==1.2.1
setproctitle==1.1.10
six==1.12.0
soupsieve==1.9.5
tabulate==0.8.6
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
yarl==1.4.2
zipp==0.6.0
